--- .github/workflows/codecov.yml
@@ -1,24 +1,0 @@
-name: Codecov
-on: [push, pull_request]
-jobs:
-  run:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Fetch
-        uses: actions/checkout@master
-      - name: Test
-        run: |
-          pip install -r requirements.txt
-          pip install pandas sphinx_rtd_theme
-          pip install pandas pandas
-          pip install pytest
-          pip install pytest-cov
-          pip install dvc
-          pip install pydrive2
-          pip install .
-          pip install coverage
-          pytest --cov-report xml --cov=ITMO_FS || true
-      - name: Upload
-        uses: codecov/codecov-action@v1
-        with:
-          files: .coverage
