+++ .gitignore
@@ -1,5 +1,8 @@
 .idea
+.venv
 ITMO_FS.egg-info
 dist
 build
 common
+.coverage
+coverage.xml
