+++ CITATION.cff
@@ -1,0 +1,18 @@
+cff-version: 1.2.0
+message: "If you use ITMO_FS in your work, please cite it using the following metadata."
+title: "ITMO_FS: Feature Selection Library"
+version: "0.3.3"
+date-released: "2025-07-31"
+url: "https://pypi.org/project/itmo_fs/"
+repository-code: "https://github.com/ITMO-FS/ITMO_FS"
+license: "MIT"
+type: software
+authors:
+
+family-names: "Pilnensky"
+given-names: "Nikita"
+affiliation: "Independent developer (ITMO University)"
+
+family-names: "Smetannikov"
+given-names: "Ivan"
+affiliation: "Independent researcher (PhD)"
