+++ ITMO_FS/__about__.py
@@ -3,4 +3,4 @@ __all__ = ["__title__", "__uri__", "__version__"]
 __title__ = "ITMO_FS"
 __uri__ = "https://github.com/ctlab/ITMO_FS"
 
-__version__ = "0.3.4"
+__version__ = "0.3.5"
