+++ requirements.txt
@@ -1,5 +1,5 @@
 imbalanced-learn
-numpy~=1.19.2
+numpy~=1.22
 scipy~=1.5.2
 scikit-learn~=0.24
 qpsolvers
