--- setup.cfg
@@ -1,8 +1,0 @@
-[metadata]
-description-file = README.rst
-
-[aliases]
-test = pytest
-
-[tool:pytest]
-addopts = --doctest-modules
